+++ magenpy/AnnotationMatrix.py
@@ -1,52 +1,53 @@
 import numpy as np
-import pandas as pd
-from magenpy.utils.compute_utils import intersect_arrays
 
 
 class AnnotationMatrix(object):
 
-    def __init__(self):
+    def __init__(self, annotation_table=None, annotations=None):
+        """
+        Initialize an AnnotationMatrix object
 
-        self._annot_matrix = None
-        self._snps = None
-        self._chromosome = None
-        self._annotations = None
-        self._binary_annotations = None
+        :param annotation_table: A pandas dataframe containing the annotation information.
+        :param annotations: A list of array of columns to consider as annotations. If not provided, will be
+        inferred heuristically, though we recommend that the user specify this information.
+        """
+
+        self.table = annotation_table
+        self._annotations = annotations
+
+        if self.table is not None:
+            if self._annotations is None:
+                self._annotations = [ann for ann in self.table.columns if ann not in ('CHR', 'SNP', 'POS')]
+                if len(self._annotations) < 1:
+                    self._annotations = None
 
     @classmethod
-    def from_file(cls, annot_file, **read_csv_kwargs):
+    def from_file(cls, annot_file, annot_format='magenpy', annot_parser=None,
+                  **parse_kwargs):
         """
-        Takes an LDSC-style annotation file `annot_file` and reads the table
-        into a pandas dataframe.
-        NOTE: Assumes that SNPs are sorted by BP within a chromosome.
+        Takes an annotation file and initializes an annotation matrix object from it.
 
-        :param annot_file: LDSC-style annotation file, containing columns for
-        the chromosome (CHR), SNP ID (SNP), Base pair position (BP), position in centi Morgans (CM),
-        and the annotations to load into the model.
-
-        :param read_csv_kwargs: Keyword arguments to pass to pandas `read_csv`.
+        :param annot_file: The path to the annotation file.
+        :param annot_format: The format of the annotation file. For now, we mainly support
+        annotation files in the `magenpy` and `ldsc` formats.
+        :param annot_parser: An `AnnotationMatrixParser` derived object, which can be tailored to
+        specific annotation formats that the user has.
+        :param parse_kwargs: arguments for the pandas `read_csv` function, such as the delimiter.
         """
 
-        annot_mat = cls()
+        from .parsers.annotation_parsers import AnnotationMatrixParser, LDSCAnnotationMatrixParser
 
-        # If the delimiter is not specified, assume whitespace by default:
-        if 'sep' not in read_csv_kwargs and 'delimiter' not in read_csv_kwargs:
-            read_csv_kwargs['delim_whitespace'] = True
+        if annot_parser is None:
+            if annot_format == 'magenpy':
+                annot_parser = AnnotationMatrixParser(None, **parse_kwargs)
+            elif annot_format == 'ldsc':
+                annot_parser = LDSCAnnotationMatrixParser(None, **parse_kwargs)
+            else:
+                raise KeyError(f"Annotation matrix format {annot_format} not recognized!")
 
-        try:
-            annot_df = pd.read_csv(annot_file, **read_csv_kwargs)
-        except Exception as e:
-            raise e
+        annot_table, annotations = annot_parser.parse(annot_file)
 
-        annot_mat._snps = annot_df['SNP'].values
-        annot_mat._chromosome = annot_df['CHR'][0]
-
-        annot_df = annot_df.drop(['CHR', 'SNP', 'BP', 'CM', 'base'], axis=1)
-
-        annot_mat._annotations = np.array(annot_df.columns)
-        annot_mat._binary_annotations = np.array([c for c in annot_df.columns if len(annot_df[c].unique()) == 2])
-
-        annot_mat._annot_matrix = annot_df.values
+        annot_mat = cls(annotation_table=annot_table, annotations=annotations)
 
         return annot_mat
 
@@ -56,40 +57,53 @@ class AnnotationMatrix(object):
 
     @property
     def n_snps(self):
-        return len(self.snps)
+        return len(self.table)
 
     @property
     def chromosome(self):
-        return self._chromosome
+        chrom = self.chromosomes
+        if chrom is not None:
+            if len(chrom) == 1:
+                return chrom[0]
+
+    @property
+    def chromosomes(self):
+        if 'CHR' in self.table.columns:
+            return self.table['CHR'].unique()
 
     @property
     def snps(self):
-        return self._snps
+        return self.table['SNP'].values
 
     @property
     def n_annotations(self):
-        return len(self.annotations)
+        if self.annotations is None:
+            return 0
+        else:
+            return len(self.annotations)
 
     @property
     def binary_annotations(self):
-        return self._binary_annotations
+        assert self.annotations is not None
+        return np.array([c for c in self.annotations
+                         if len(self.table[c].unique()) == 2])
 
     @property
     def annotations(self):
         return self._annotations
-
-    def annotation_index(self, annot):
-        return list(self.annotations).index(annot)
 
     def values(self, add_intercept=False):
         """
         Returns the annotation matrix.
         :param add_intercept: Adds a base annotation corresponding to the intercept.
         """
+        if self.annotations is None:
+            raise KeyError("No annotations are defined in this table!")
+        annot_mat = self.table[self.annotations].values
         if add_intercept:
-            return np.hstack([np.ones((self._annot_matrix.shape[0], 1)), self._annot_matrix])
+            return np.hstack([np.ones((annot_mat.shape[0], 1)), annot_mat])
         else:
-            return self._annot_matrix
+            return annot_mat
 
     def filter_snps(self, extract_snps=None, extract_file=None):
         """
@@ -107,10 +121,11 @@ class AnnotationMatrix(object):
             from .parsers.misc_parsers import read_snp_filter_file
             extract_snps = read_snp_filter_file(extract_file)
 
-        arr_idx = intersect_arrays(self._snps, extract_snps, return_index=True)
+        from magenpy.utils.compute_utils import intersect_arrays
 
-        self._snps = self._snps[arr_idx]
-        self._annot_matrix = self._annot_matrix[arr_idx, :]
+        arr_idx = intersect_arrays(self.snps, extract_snps, return_index=True)
+
+        self.table = self.table.iloc[arr_idx, :].reset_index()
 
     def filter_annotations(self, keep_annotations):
         """
@@ -118,16 +133,127 @@ class AnnotationMatrix(object):
         :param keep_annotations: A list or vector of annotations to keep.
         """
 
-        ann_idx = intersect_arrays(self._annotations, keep_annotations, return_index=True)
+        if self.annotations is None:
+            return
 
-        self._annotations = self._annotations[ann_idx]
-        self._annot_matrix = self._annot_matrix[:, ann_idx]
+        self._annotations = [annot for annot in self._annotations if annot in keep_annotations]
+        self.table = self.table[['CHR', 'SNP', 'POS'] + self._annotations]
 
-        self._binary_annotations = intersect_arrays(self._annotations, self._binary_annotations)
+    def add_annotation(self, annot_vec, annotation_name):
+        """
+        Add an annotation vector or list to the AnnotationMatrix object.
+        :param annot_vec: A vector/list/Series containing the annotation information for each SNP in the
+        AnnotationMatrix. For now, it's the responsibility of the user to make sure that the annotation
+        list or vector are sorted properly.
+        :param annotation_name: The name of the annotation to create. Make sure the name is not already
+        in the matrix!
+        """
+
+        if self.annotations is not None:
+            assert annotation_name not in self.annotations
+        assert len(annot_vec) == self.n_snps
+
+        self.table[annotation_name] = annot_vec
+
+        if self.annotations is None:
+            self._annotations = [annotation_name]
+        else:
+            self._annotations = list(self._annotations) + [annotation_name]
+
+    def add_annotation_from_bed(self, bed_file, annotation_name):
+        """
+        Add an annotation to the AnnotationMatrix from a BED file that lists
+        the range of coordinates associated with that annotation (e.g. coding regions, enhancers, etc.).
+        The BED file has to adhere to the format specified by,
+        https://uswest.ensembl.org/info/website/upload/bed.html
+        with the first three columns being:
+
+        CHR StartCoordinate EndCoordinate ...
+
+        NOTE: This implementation is quite slow at the moment. May need to find more efficient
+        ways to do the merge over list of ranges.
+
+        :param bed_file: The path to the BED file containing the annotation coordinates.
+        :param annotation_name: The name of the annotation to create. Make sure the name is not already
+        in the matrix!
+        """
+
+        from .parsers.annotation_parsers import parse_annotation_bed_file
+
+        if self.annotations is not None:
+            assert annotation_name not in self.annotations
+
+        bed_df = parse_annotation_bed_file(bed_file)
+        # Group the BED annotation file by chromosome:
+        range_groups = bed_df.groupby('CHR').groups
+
+        def annotation_overlap(row):
+            """
+            This function takes a row from the annotation matrix table
+            and returns True if and only if the BP position for the
+            SNP is within the range specified by the annotation BED file.
+            """
+            try:
+                chr_range = bed_df.iloc[range_groups[row['CHR']], :]
+            except KeyError:
+                return False
+
+            check = (chr_range.Start <= row['POS']) & (chr_range.End >= row['POS'])
+            return int(check.any())
+
+        self.table[annotation_name] = self.table.apply(annotation_overlap, axis=1)
+
+        if self.annotations is None:
+            self._annotations = [annotation_name]
+        else:
+            self._annotations = list(self._annotations) + [annotation_name]
 
     def get_binary_annotation_index(self, bin_annot):
         """
         Get the indices of all SNPs that belong to binary annotation `bin_annot`
         """
-        assert bin_annot in self._binary_annotations
-        return np.where(self._annot_matrix[:, self.annotation_index(bin_annot)] == 1)[0]
+        assert bin_annot in self.binary_annotations
+        return np.where(self.table[bin_annot] == 1)[0]
+
+    def split_by_chromosome(self):
+        """
+        Split the annotation matrix by chromosome, so that we would
+        have a separate `AnnotationMatrix` object for each chromosome.
+        """
+
+        if 'CHR' in self.table.columns:
+            chrom_tables = self.table.groupby('CHR')
+            return {
+                c: AnnotationMatrix(annotation_table=chrom_tables.get_group(c),
+                                    annotations=self.annotations)
+                for c in chrom_tables.groups
+            }
+        else:
+            raise KeyError("Chromosome information is not available in the annotation table!")
+
+    def to_file(self, output_path, col_subset=None, compress=True, **to_csv_kwargs):
+        """
+        Write the annotation matrix to file.
+
+        :param output_path: The path and prefix to the file where to write the annotation matrix.
+        :param col_subset: A subset of the columns to write to file.
+        :param compress: Whether to compress the output file (default: True).
+        :param to_csv_kwargs: Key-word arguments to the pandas csv writer.
+        """
+
+        if 'sep' not in to_csv_kwargs and 'delimiter' not in to_csv_kwargs:
+            to_csv_kwargs['sep'] = '\t'
+
+        if 'index' not in to_csv_kwargs:
+            to_csv_kwargs['index'] = False
+
+        if col_subset is not None:
+            table = self.table[col_subset]
+        else:
+            table = self.table
+
+        file_name = output_path + '.annot'
+        if compress:
+            file_name += '.gz'
+
+        table.to_csv(file_name, **to_csv_kwargs)